pyproject.toml
[project]
name = "stock-screener"
version = "0.1.0"
description = "株価期待値スクリーニングツール - TSE Prime market stock screening by expected value"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=0.2",
    "pandas>=2.0",
    "numpy>=1.26",
    "streamlit>=1.30",
    "pyyaml>=6.0",
    "xlrd>=2.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow: marks tests as slow (require network access)"]
